+++ examples/run_local_in_safari_via_set_config_driver/test_todomvc.py
@@ -27,8 +27,7 @@ from selenium import webdriver
 
 
 def test_completes_todo():
-    webdriver.Op
-    # browser.config.driver = webdriver.Safari(service=SafariService())
+    browser.config.driver = webdriver.Safari(service=SafariService())
 
     browser.open('http://todomvc.com/examples/emberjs/')
     browser.should(have.title_containing('TodoMVC'))