+++ tests/base_test.py
@@ -24,6 +24,7 @@ from selenium.webdriver.chrome.service import Service
 from webdriver_manager.chrome import ChromeDriverManager
 from webdriver_manager.core.utils import ChromeType
 
+from selene import support
 from selene.support.shared import browser
 
 
@@ -31,7 +32,9 @@ class BaseTest:
     def setup_method(self):
         browser.config.driver = webdriver.Chrome(
             service=Service(
-                ChromeDriverManager(chrome_type=ChromeType.GOOGLE).install()
+                support._extensions.webdriver_manager.patch._to_find_chromedrivers_from_115(
+                    ChromeDriverManager(chrome_type=ChromeType.GOOGLE)
+                ).install()
             )
         )