+++ tests/integration/save_page_source_test.py
@@ -28,7 +28,7 @@ from selenium.webdriver.chrome.service import Service
 from webdriver_manager.chrome import ChromeDriverManager  # type: ignore
 from webdriver_manager.core.utils import ChromeType  # type: ignore
 
-from selene import have, be, browser, Browser, Config, command
+from selene import have, be, browser, Browser, Config, command, support
 from selene.core import query
 from selene.core.exceptions import TimeoutException
 from tests import resources
@@ -40,7 +40,11 @@ EMPTY_PAGE_URL = resources.url('empty.html')
 @pytest.fixture(scope='module')
 def the_driver():
     driver = webdriver.Chrome(
-        service=Service(ChromeDriverManager(chrome_type=ChromeType.GOOGLE).install()),
+        service=Service(
+            support._extensions.webdriver_manager.patch._to_find_chromedrivers_from_115(
+                ChromeDriverManager(chrome_type=ChromeType.GOOGLE)
+            ).install()
+        ),
         options=headless_chrome_options(),
     )